requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
pandas==2.2.0
numpy==1.26.4
torch==2.2.0
joblib==1.3.2
scikit-learn==1.4.0
python-dotenv==1.0.0
